move session edit related components to Edit directory
