+++ client/src/SessionView/Edit/EditPasswordDialog.js
@@ -10,7 +10,7 @@ import {
 } from '@material-ui/core'
 import { Trans } from 'react-i18next'
 
-import { useSessionContext } from '../SessionProvider'
+import { useSessionContext } from '../useSessionContext'
 
 import { usePassword } from './queries'
