+++ client/src/SessionView/Edit/index.js
@@ -1,1 +1,4 @@
 export * from './useEdit'
+export * from './EditPrompt'
+export * from './EditPromptProvider'
+export * from './EditButton'
