--- client/src/SessionView/EditButton/index.js
@@ -1,2 +1,0 @@
-export { EditButton } from './EditButton'
-export * from './EditPrompt'
