+++ client/src/SessionView/useSessionContext.js
@@ -1,0 +1,4 @@
+import React, { useContext } from 'react'
+
+export const SessionContext = React.createContext()
+export const useSessionContext = () => useContext(SessionContext)
